Refactor AppState God Object into focused state managers
Split the 821-line AppState class into focused, single-responsibility managers:

**New Structure:**
- `types.ts` (139 lines): Shared type definitions and enums
- `terminal-state-manager.ts` (342 lines): Terminal CRUD, primary selection, ordering
- `workspace-state-manager.ts` (105 lines): Workspace path validation and persistence
- `state.ts` (388 lines): Composition, coordination, and backward compatibility

**Benefits:**
- ✅ Single Responsibility: Each manager has one clear purpose
- ✅ Testability: Can mock just the manager you need
- ✅ Maintainability: Changes to one concern don't risk breaking others
- ✅ Understandability: Developers only need to understand relevant manager
- ✅ Reduced Conflicts: Changes distributed across multiple files
- ✅ Backward Compatible: All existing code continues to work via delegation

**Changes:**
- Extracted terminal state management into TerminalStateManager
- Extracted workspace state management into WorkspaceStateManager
- Moved all type definitions to dedicated types.ts file
- AppState now composes managers instead of managing everything directly
- Added @deprecated tags to legacy methods that delegate to managers
- Maintained full backward compatibility - no breaking changes

Closes #598
